+++ src/components/teamInput.js
@@ -1,12 +1,35 @@
+import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
+import { useState, useEffect } from 'react';
 
 import InputGroup from 'react-bootstrap/InputGroup';
 import FormControl from 'react-bootstrap/FormControl';
 
-const TeamInput = ({ addTeam, setTeam, teamName }) => {
+import { getTeams, getGames } from '../redux/actions';
+import { addTeamToStorage } from '../helpers/gameHelpers';
+
+const TeamInput = ({ dispatch }) => {
+  const [teamName, setTeamName] = useState('');
+
+  useEffect(() => {
+    dispatch(getTeams());
+    dispatch(getGames());
+  }, []);
+
+  const addTeamToDisplay = () => {
+    addTeamToStorage(teamName);
+    setTeamName('');
+    dispatch(getTeams());
+    dispatch(getGames());
+  };
+
+  const changeTeamName = (e) => {
+    setTeamName(e.target.value);
+  };
+
   const handleKeyPress = (e) => {
     if (e.key === 'Enter' || e.keycode === 13) {
-      addTeam();
+      addTeamToDisplay();
     }
   };
   return (
@@ -16,25 +39,18 @@ const TeamInput = ({ addTeam, setTeam, teamName }) => {
         aria-label="New Team"
         aria-describedby="team-input"
         value={teamName}
-        onChange={setTeam}
+        onChange={changeTeamName}
         onKeyPress={handleKeyPress}
       />
       <InputGroup.Append>
-        <InputGroup.Text id="add-btn" onClick={addTeam}>Add Team</InputGroup.Text>
+        <InputGroup.Text id="add-btn" onClick={addTeamToDisplay}>Add Team</InputGroup.Text>
       </InputGroup.Append>
     </InputGroup>
   );
 };
 
 TeamInput.propTypes = {
-  addTeam: PropTypes.func.isRequired,
-  setTeam: PropTypes.func.isRequired,
-  teamName: PropTypes.string,
-
+  dispatch: PropTypes.func.isRequired,
 };
 
-TeamInput.defaultProps = {
-  teamName: '',
-};
-
-export default (TeamInput);
+export default connect(null)(TeamInput);
